+++ packages/crypto/src/index.browser.ts
@@ -1,0 +1,12 @@
+import cryptoApi from './browser';
+
+export * from './types';
+
+export const {
+  bufferFromString,
+  decrypt,
+  encrypt,
+  keyFromPassword,
+  randomBytes,
+  stringFromBuffer,
+} = cryptoApi;