+++ packages/fuel-gauge/src/predicate/predicate-arguments.test.ts
@@ -6,6 +6,8 @@ import predicateBytesMainArgsStruct from '../../fixtures/forc-projects/predicate
 import predicateAbiMainArgsStruct from '../../fixtures/forc-projects/predicate-main-args-struct/out/debug/predicate-main-args-struct-abi.json';
 import predicateBytesMainArgsVector from '../../fixtures/forc-projects/predicate-main-args-vector';
 import predicateAbiMainArgsVector from '../../fixtures/forc-projects/predicate-main-args-vector/out/debug/predicate-main-args-vector-abi.json';
+import predicateBytesMulti from '../../fixtures/forc-projects/predicate-multi-args';
+import predicateAbiMulti from '../../fixtures/forc-projects/predicate-multi-args/out/debug/predicate-multi-args-abi.json';
 import predicateBytesStruct from '../../fixtures/forc-projects/predicate-struct';
 import predicateBytesU32 from '../../fixtures/forc-projects/predicate-u32';
 import type { Validation } from '../types/predicate';
@@ -286,6 +288,8 @@ describe('Predicate', () => {
       const initialPredicateBalance = await fundPredicate(wallet, predicate, amountToPredicate);
       const initialReceiverBalance = await receiver.getBalance();
 
+      // #region predicate-struct-arg
+      // #context const predicate = new Predicate(bytecode, chainId, abi);
       const tx = await predicate
         .setData({
           has_account: true,
@@ -293,6 +297,7 @@ describe('Predicate', () => {
         })
         .transfer(receiver.address, amountToReceiver);
       await tx.waitForResult();
+      // #endregion predicate-struct-arg
 
       await assertBalances(
         predicate,
@@ -349,6 +354,45 @@ describe('Predicate', () => {
         initialReceiverBalance,
         amountToPredicate,
         amountToReceiver
+      );
+    });
+
+    it('calls a predicate with valid multiple arguments and returns true', async () => {
+      const amountToPredicate = 100;
+      const amountToReceiver = 50;
+      const predicate = new Predicate(predicateBytesMulti, chainId, predicateAbiMulti);
+
+      const initialPredicateBalance = await fundPredicate(wallet, predicate, amountToPredicate);
+      const initialReceiverBalance = await receiver.getBalance();
+
+      // #region predicate-multi-args
+      // #context const predicate = new Predicate(bytecode, chainId, abi);
+      predicate.setData(20, 30);
+      const tx = await predicate.transfer(receiver.address, amountToReceiver);
+      await tx.waitForResult();
+      // #endregion predicate-multi-args
+
+      await assertBalances(
+        predicate,
+        receiver,
+        initialPredicateBalance,
+        initialReceiverBalance,
+        amountToPredicate,
+        amountToReceiver
+      );
+    });
+
+    it('calls a predicate with invalid multiple arguments and returns false', async () => {
+      const amountToPredicate = 100;
+      const predicate = new Predicate(predicateBytesMulti, chainId, predicateAbiMulti);
+
+      const initialPredicateBalance = await fundPredicate(wallet, predicate, amountToPredicate);
+
+      // Check the UTXOs locked with the predicate hash
+      expect(toNumber(initialPredicateBalance)).toBeGreaterThanOrEqual(amountToPredicate);
+
+      await expect(predicate.setData(20, 20).transfer(receiver.address, 50)).rejects.toThrow(
+        'Invalid transaction'
       );
     });
   });