+++ packages/versions/src/lib/getSupportedVersions.ts
@@ -2,6 +2,6 @@ export function getSupportedVersions() {
   return {
     FORC: '0.40.1',
     FUEL_CORE: '0.19.1',
-    FUELS: '0.48.2',
+    FUELS: '0.49.0',
   };
 }