+++ packages/providers/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 0.49.0
+
+### Minor Changes
+
+- rename package keystore to crypto, by [@Torres-ssf](https://github.com/Torres-ssf) (See [#1140](https://github.com/FuelLabs/fuels-ts/pull/1140))
+
+### Patch Changes
+
+- 🐞 Fix predicate check on estimateTxDependencies, by [@luizstacio](https://github.com/luizstacio) (See [#1142](https://github.com/FuelLabs/fuels-ts/pull/1142))
+
 ## 0.48.2
 
 ### Patch Changes